refactor: 책 검색 기능 한 파일로 합침
- 코드 구조 개선을 위해 관련 기능 하나의 파일로 합침
