--- src/routes/searchBookRoutes.ts
@@ -1,8 +1,0 @@
-import { Router } from "express";
-import { searchBooks } from "../controllers/searchBookController";
-
-const router = Router();
-
-router.get('/', searchBooks);
-
-export default router;
