+++ src/server.ts
@@ -1,7 +1,7 @@
-import app from './app';
+import app from "./app";
 
 const port = process.env.PORT;
 
 app.listen(port, () => {
-    console.log(`Server is running on port ${port}`)
-})
+  console.log(`Server is running on port ${port}`);
+});
